+++ CHANGELOG.md
@@ -62,5 +62,5 @@
 #### ⟳ Changelog Beta version 0.1.8 (2026-02-07)
 - Add new theme Amber Dusk.
 - Modify list theme.    
-- Add button to open terminal in the project folder.
+- Add button to open terminal.
 - New version 0.1.9 (2026-02-07)