+++ CHANGELOG.md
@@ -68,3 +68,5 @@
 #### ⟳ Changelog Beta version 0.1.10 (2026-02-08)
 - Fixed color of the text in the notes tab.
 - Changed info layout in settings tab.
+- Add copy icon in notes tab.
+